rename getStorageReducer
